fix: add throws ResourceNotFoundException to Course DAO tests
